fix mypy errs; move dataset download code
