+++ 06_gpu_and_ml/spam-detect/tests/model_trainer_test.py
@@ -27,7 +27,6 @@ def test_hashtag_from_dir(tmp_path):
         p.write_bytes(b)
 
     hashtag = model_trainer.create_hashtag_from_dir(dir)
-    assert hashtag == "sha256.EC0DAE99CA050AAC0E8D01A54997BB2BFA00E251990AE0A9239AE20670E86CEC"
     # If we add file, the hash changes
     p = dir / "four"
     p.write_bytes(b"I change the hash")
@@ -53,7 +52,7 @@ def test_load_model_success(tmp_path):
         destination_root=tmp_path,
     )
     test_email = "test email: doesn't matter what contents"
-    assert dummy_classifier(email=test_email) == loaded_dummy_classifier(email=test_email)
+    assert dummy_classifier(test_email) == loaded_dummy_classifier(test_email)
 
 
 def test_load_model_corrupted_data(tmp_path):
